Push #6
The swipe directions work, the trail is scuffed but not by much so i have made it invisible for now. 
all that needs to be done is to build and run on android device and also to fix bugs/ add some extras if i feel like it.

but i am ultimately done.